Move controller start to subcommand in preparation of the `render` command.
